% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MASSICCC_plotClasses.R
\name{clusvisMixtcomp}
\alias{clusvisMixtcomp}
\title{This function estimates the parameters used for visualization of model-based clustering performs with R package ClusVis 
To achieve the parameter infernece, it automatically samples probabilities of classification from the model parameters}
\usage{
clusvisMixtcomp(resmixtcomp, sample.size = 5000, maxit = 10^3,
  nbrandomInit = 12, nbcpu = 1)
}
\arguments{
\item{sample.size}{numeric. Number of probabilities of classification sampled for parameter inference.}

\item{maxit}{numeric. It limits the number of iterations for the Quasi-Newton algorithm (default 1000).}

\item{nbrandomInit}{numeric. It defines the number of random initialization of the Quasi-Newton algorithm.}

\item{nbcpu}{numeric. It specifies the number of CPU (only for linux, R package parallel must by loaded).}

\item{resmixcomp}{list. It is an object return by MixComp.}
}
\description{
This function estimates the parameters used for visualization of model-based clustering performs with R package ClusVis 
To achieve the parameter infernece, it automatically samples probabilities of classification from the model parameters
}
