% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTCOMP_misc.R
\name{computeLogTik}
\alias{computeLogTik}
\title{Compute log(tik)}
\usage{
computeLogTik(outMixtComp)
}
\arguments{
\item{outMixtComp}{output object of \link{mixtCompCluster} or \link{mixtCompPredict} functions.}
}
\value{
matrix with tik
}
\description{
Compute the log(tik)
}
\examples{
\dontrun{
# path to files
pathToData <- system.file("extdata", "data.csv", package = "RMixtComp")
pathToDescriptor <- system.file("extdata", "descUnsupervised.csv", package = "RMixtComp")

resGetData <- getData(c(pathToData, pathToDescriptor))

# define the algorithm's parameters
mcStrategy <- createMcStrategy()

# run RMixtCompt for clustering
res <- mixtCompCluster(resGetData$lm, mcStrategy, nbClass = 2, confidenceLevel = 0.95)

logTik <- computeLogTik(res)
}

}
