% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTCOMP_formatFunctions.R
\name{createFunctional}
\alias{createFunctional}
\title{Create a functional in MixtComp format}
\usage{
createFunctional(time, value)
}
\arguments{
\item{time}{vector containing the time of the functional}

\item{value}{vector containing the value of the functional}
}
\value{
The functional data formatted to the mixtcomp standard
}
\description{
Create a functional in MixtComp format
}
\examples{
mat <- matrix(c(1, 2, 3, 9, 1, 1.5, 15, 1000), ncol = 2)
createFunctional(mat[,1], mat[,2])


}
