% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GENDATA_functionalInterPolyGenerator.R
\name{functionalInterPolyGenerator}
\alias{functionalInterPolyGenerator}
\title{Generate a functional observation using an interpolating polynomial instead of the generating model.
This allows for a little more flexibility and convenience in specifying data than functionalGenerator
which is currently bugged anyway.}
\usage{
functionalInterPolyGenerator(present, param)
}
\arguments{
\item{param}{A structure describing the points by which the interpolating polynomial is supposed to go through.}
}
\value{
The generated observation as a string.
}
\description{
Generate a functional observation using an interpolating polynomial instead of the generating model.
This allows for a little more flexibility and convenience in specifying data than functionalGenerator
which is currently bugged anyway.
}
\examples{
param <- list(
  x = c(0., 10., 20.),
  y = c(0., 10., 0.),
  sd = 0.1,
  tMin = 0.,
  tMax = 20.,
  nTime = 100)
functionalInterPolyGenerator(param)
}
