% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTCOMP_getter.R
\name{getBIC}
\alias{getBIC}
\alias{getICL}
\title{Get criterion value}
\usage{
getBIC(outMixtComp)

getICL(outMixtComp)
}
\arguments{
\item{outMixtComp}{object of class \emph{MixtCompLearn} or \emph{MixtComp} obtained using \link{mixtCompLearn} or \link{mixtCompPredict} functions.}
}
\value{
value of the criterion
}
\description{
Get criterion value
}
\examples{
\donttest{
data(simData)
 
# define the algorithm's parameters
algo <- list(nbBurnInIter = 100,
             nbIter = 100,
             nbGibbsBurnInIter = 50,
             nbGibbsIter = 50,
             nInitPerClass = 10,
             nSemTry = 20,
             confidenceLevel = 0.95)

# run RMixtCompt for clustering
resLearn <- mixtCompLearn(simDataLearn$matrix, desc$unsupervised, algo, nbClass = 2)

# get criterion
bic <- getBIC(resLearn)
icl <- getICL(resLearn)
}

}
\seealso{
Other getter: \code{\link{getCompletedData}},
  \code{\link{getEmpiricTik}}, \code{\link{getParam}},
  \code{\link{getPartition}}, \code{\link{getType}}
}
\concept{getter}
