% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTCOMP_getParam.R
\name{getParam}
\alias{getParam}
\title{Get the estimated parameter}
\usage{
getParam(outMixtComp, varName)
}
\arguments{
\item{outMixtComp}{output object of \link{mixtCompCluster} or \link{mixtCompPredict} functions.}

\item{varName}{name of the variable to get parameter}
}
\value{
the parameter of the variable
}
\description{
Get the estimated parameter
}
\examples{
\dontrun{
# path to files
pathToData <- system.file("extdata", "data.csv", package = "RMixtComp")
pathToDescriptor <- system.file("extdata", "descUnsupervised.csv", package = "RMixtComp")

resGetData <- getData(c(pathToData, pathToDescriptor))


# define the algorithm's parameters
mcStrategy <- createMcStrategy()

# run RMixtCompt for clustering
res <- mixtCompCluster(resGetData$lm, mcStrategy, nbClass = 2, confidenceLevel = 0.95)

# get param
param <- getParam(res, "poisson1")
}

}
