% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTCOMP_getter.R
\name{getPartition}
\alias{getPartition}
\title{Get the estimated class from MixtComp object}
\usage{
getPartition(outMixtComp)
}
\arguments{
\item{outMixtComp}{object of class \emph{MixtCompLearn} or \emph{MixtComp} obtained using \link{mixtCompLearn} or \link{mixtCompPredict} functions.}
}
\value{
a vector containing the estimated class for each individual.
}
\description{
Get the estimated class from MixtComp object
}
\examples{
\donttest{
data(simData)
 
# define the algorithm's parameters
algo <- list(nbBurnInIter = 100,
             nbIter = 100,
             nbGibbsBurnInIter = 50,
             nbGibbsIter = 50,
             nInitPerClass = 10,
             nSemTry = 20,
             confidenceLevel = 0.95)

# keep only 3 variables
model <- simData$model$unsupervised[c("Gaussian1", "Poisson1", "Categorical1")]

# run RMixtCompt for clustering
resLearn <- mixtCompLearn(simData$dataLearn$matrix, model, algo, nClass = 2)

# get class
estimatedClass <- getPartition(resLearn)
}

}
\seealso{
Other getter: \code{\link{getBIC}},
  \code{\link{getCompletedData}},
  \code{\link{getEmpiricTik}}, \code{\link{getParam}},
  \code{\link{getType}}
}
\author{
Quentin Grimonprez
}
\concept{getter}
