% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTCOMP_get.R
\name{getTik}
\alias{getTik}
\title{Get the empiric tik}
\usage{
getTik(outMixtcomp)
}
\arguments{
\item{outMixtComp}{output object of \link{mixtCompCluster} or \link{mixtCompPredict} fucntions.}
}
\value{
a matrix containing the tik for each individuals (in row) and each class (in column).
}
\description{
Get the empiric tik
}
\examples{
\dontrun{
# load the data
resGetData <- getData(c("dataTrain.csv", "descriptor.csv")) 
resGetNewData <- getData(c("dataTest.csv", "descriptor.csv")) 

# define the algorithm's parameters
mcStrategy <- list(nbBurnInIter = 100,
nbIter = 100,
nbGibbsBurnInIter = 50,
nbGibbsIter = 50,
parameterEdgeAuthorized = FALSE)

# run RMixtCompt for clustering
res <- mixtCompCluster(resGetData$lm, mcStrategy, nbClass = 2, confidenceLevel = 0.95)

# get tik
tik <- getTik(res)
}

}
