% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTCOMP_getter.R
\name{getType}
\alias{getType}
\title{Get the type of model for each variable}
\usage{
getType(outMixtComp, with.z_class = FALSE)
}
\arguments{
\item{outMixtComp}{output object of \link{mixtCompCluster} or \link{mixtCompPredict} fucntions.}

\item{with.z_class}{if TRUE, the type of z_class is returned.}
}
\value{
a vector containing the type of models associated with each individual.
}
\description{
Get the type of model for each variable
}
\examples{
\dontrun{
# path to files
pathToData <- system.file("extdata", "data.csv", package = "RMixtComp")
pathToDescriptor <- system.file("extdata", "descriptor.csv", package = "RMixtComp")

resGetData <- getData(c(pathToData, pathToDescriptor))


# define the algorithm's parameters
mcStrategy <- list(nbBurnInIter = 100,
                   nbIter = 100,
                   nbGibbsBurnInIter = 50,
                   nbGibbsIter = 50,
                   ratioInitialization = 0.1)

# run RMixtCompt for clustering
res <- mixtCompCluster(resGetData$lm, mcStrategy, nbClass = 2, confidenceLevel = 0.95)

# get type
type <- getType(res)
}
}
