% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTCOMP_get.R
\name{getZ_class}
\alias{getZ_class}
\title{Get the estimated class from MixtComp object}
\usage{
getZ_class(outMixtcomp)
}
\arguments{
\item{outMixtComp}{output object of \link{mixtCompCluster} or \link{mixtCompPredict} fucntions.}
}
\value{
a vector containing the estimated class for each individual.
}
\description{
Get the estimated class from MixtComp object
}
\examples{
\dontrun{
# load the data
resGetData <- getData(c("dataTrain.csv", "descriptor.csv")) 
resGetNewData <- getData(c("dataTest.csv", "descriptor.csv")) 

# define the algorithm's parameters
mcStrategy <- list(nbBurnInIter = 100,
nbIter = 100,
nbGibbsBurnInIter = 50,
nbGibbsIter = 50,
parameterEdgeAuthorized = FALSE)

# run RMixtCompt for clustering
res <- mixtCompCluster(resGetData$lm, mcStrategy, nbClass = 2, confidenceLevel = 0.95)

# get class
estiamtedClass <- getZ_class(res)
}

}
