% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT_plotfunctions.R
\name{heatmapTikSorted}
\alias{heatmapTikSorted}
\title{Heatmap of the tik = P(Z_i=k|x_i)}
\usage{
heatmapTikSorted(output, ...)
}
\arguments{
\item{output}{object returned by function \emph{mixtCompCluster}}

\item{...}{arguments to be passed to plot_ly}
}
\description{
Heatmap of the tik = P(Z_i=k|x_i)
}
\details{
Observation are sorted according to the hard partition then for each component
they are sorted by decreasing order of their tik's
}
\examples{
\donttest{
library(RMixtComp)

data(simData)
 
# define the algorithm's parameters
algo <- createAlgo()

# run RMixtCompt in unsupervised clustering mode + data as matrix
res <- mixtCompLearn(simDataLearn$matrix, desc$unsupervised, algo, nbClass = 2:4)

# plot
heatmapTikSorted(res)

}

}
\seealso{
getTik

Other plot: \code{\link{heatmapClass}},
  \code{\link{heatmapVbles}}, \code{\link{histMisclassif}},
  \code{\link{plotConvergence}},
  \code{\link{plotDataBoxplot}}, \code{\link{plotDataCI}},
  \code{\link{plotDiscrimClass}},
  \code{\link{plotDiscrimVbles}}
}
\author{
Matthieu MARBAC
}
\concept{plot}
