% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTCOMP_methods.R
\name{plot.MixtComp}
\alias{plot.MixtComp}
\title{Plot of a \code{\link{MixtComp}} object}
\usage{
\method{plot}{MixtComp}(x, nVarMaxToPlot = 5, pkg = c("ggplot2",
  "plotly"), ...)
}
\arguments{
\item{x}{\code{\link{MixtComp}} object}

\item{...}{Not used.}

\item{nVarMaxToPrint}{number of variables to display}
}
\description{
Plot of a \code{\link{MixtComp}} object
}
\examples{
\donttest{
data(simData)
 
# define the algorithm's parameters
algo <- list(nbBurnInIter = 100,
             nbIter = 100,
             nbGibbsBurnInIter = 50,
             nbGibbsIter = 50,
             nInitPerClass = 10,
             nSemTry = 20,
             confidenceLevel = 0.95)

# keep only 3 variables
desc <- simDesc$unsupervised[c("Gaussian1", "Poisson1", "Categorical1")]

# run RMixtCompt in unsupervised clustering mode + data as matrix
resLearn <- mixtCompLearn(simDataLearn$matrix, desc, algo, nClass = 2:4)

plot(resLearn$res[[1]])
}
}
\seealso{
\link{mixtCompLearn} \link{mixtCompPredict}

Other plot: \code{\link{heatmapClass}},
  \code{\link{heatmapTikSorted}}, \code{\link{heatmapVar}},
  \code{\link{histMisclassif}},
  \code{\link{plotConvergence}},
  \code{\link{plotDataBoxplot}}, \code{\link{plotDataCI}},
  \code{\link{plotDiscrimClass}},
  \code{\link{plotDiscrimVar}},
  \code{\link{plotProportion}}
}
\concept{plot}
