% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MASSICCC_plotClasses.R
\name{plotComponentClusVis}
\alias{plotComponentClusVis}
\title{This function is a wrapper around the plotDensityClusVisu function of the ClusVisu package. It generates
a component interpretation graph using the visualization information returned by the clusvisMixtcomp function.}
\usage{
plotComponentClusVis(visInfo)
}
\arguments{
\item{visInfo}{list. The parameters estimated by the clusvisMixtcomp function.}
}
\description{
This function is a wrapper around the plotDensityClusVisu function of the ClusVisu package. It generates
a component interpretation graph using the visualization information returned by the clusvisMixtcomp function.
}
