% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTCOMP_plotUnivariateBoxplots.R
\name{plotDataBoxplot}
\alias{plotDataBoxplot}
\alias{plotFunctionalData}
\alias{plotCategoricalData}
\alias{plotBoxplotperClass}
\title{Boxplot per class}
\usage{
plotDataBoxplot(output, var, grl = TRUE, ...)
}
\arguments{
\item{output}{object returned by function \link{mixtCompCluster}}

\item{var}{name of the variable}

\item{grl}{if TRUE plot the general distribution of the data}

\item{...}{other parameters (see \emph{Details})}
}
\details{
For functional data, three other parameters are available:
\describe{
 \item{add.obs}{if TRUE, observations are added to the plot. Default = FALSE.}
 \item{ylim}{ylim of the plot.}
 \item{xlim}{xlim of the plot.}
}
}
\examples{
\dontrun{
# path to files
pathToData <- system.file("extdata", "data.csv", package = "RMixtComp")
pathToDescriptor <- system.file("extdata", "descriptor.csv", package = "RMixtComp")

resGetData <- getData(c(pathToData, pathToDescriptor))


# define the algorithm's parameters
mcStrategy <- list(nbBurnInIter = 100,
                   nbIter = 100,
                   nbGibbsBurnInIter = 50,
                   nbGibbsIter = 50,
                   ratioInitialization = 0.1)

# run RMixtCompt for clustering
res <- mixtCompCluster(resGetData$lm, mcStrategy, nbClass = 2, confidenceLevel = 0.95)

# plot
plotDataBoxplot(res, "gaussian1")

} 

}
\author{
Matthieu MARBAC
}
