% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT_plotUnivariateDistributions.R
\name{plotDataCI}
\alias{plotDataCI}
\title{Mean and 95\%-level confidence intervals per class}
\usage{
plotDataCI(output, var, class = 1:output$algo$nClass, grl = FALSE, ...)
}
\arguments{
\item{output}{object returned by function \emph{mixtCompCluster}}

\item{var}{name of the variable}

\item{class}{class to plot}

\item{grl}{if TRUE plot the CI for the dataset and not only classes}

\item{...}{other parameters (see \emph{Details})}
}
\description{
Mean and 95\%-level confidence intervals per class
}
\details{
For functional data, three other parameters are available:
\describe{
 \item{add.obs}{if TRUE, observations are added to the plot. Default = FALSE.}
 \item{add.CI}{if FALSE, confidence intervals are removed from the plot. Default = TRUE.}
 \item{xlim}{xlim of the plot.}
 \item{ylim}{ylim of the plot.}  
}
}
\examples{
\donttest{
library(RMixtComp)

data(simData)
 
# define the algorithm's parameters
algo <- createAlgo()

# run RMixtCompt in unsupervised clustering mode + data as matrix
res <- mixtCompLearn(simDataLearn$matrix, simDesc$unsupervised, algo, nClass = 2:4)
# plot
plotDataCI(res, "Gaussian1")
plotDataCI(res$res[[2]], "Poisson1")
} 

}
\seealso{
Other plot: \code{\link{heatmapClass}},
  \code{\link{heatmapTikSorted}},
  \code{\link{heatmapVbles}}, \code{\link{histMisclassif}},
  \code{\link{plotConvergence}},
  \code{\link{plotDataBoxplot}},
  \code{\link{plotDiscrimClass}},
  \code{\link{plotDiscrimVbles}}
}
\author{
Matthieu MARBAC
}
\concept{plot}
