% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT_barplot.R
\name{plotDiscrimVar}
\alias{plotDiscrimVar}
\title{Barplot of the discriminative power of the variables}
\usage{
plotDiscrimVar(output, ylim = c(0, 1), pkg = c("ggplot2", "plotly"),
  ...)
}
\arguments{
\item{output}{object returned by function \emph{mixtCompLearn}}

\item{ylim}{vector of length 2 defining the range of y-axis}

\item{pkg}{"ggplot2" or "plotly". Package used to plot}

\item{...}{arguments to be passed to plot_ly}
}
\description{
Barplot of the discriminative power of the variables
}
\details{
The discriminative power of variable j is defined by 1 - C(j)
\deqn{C(j)=  -\sum_{k=1}^K \sum_{i=1}^n P(Z_i=k|x_{ij}) ln(P(Z_i=k|x_{ij})) / (n*\log(K))}
}
\examples{
\donttest{
data(simData)
 
# define the algorithm's parameters
algo <- createAlgo()

# keep only 3 variables
model <- simData$model$unsupervised[c("Gaussian1", "Poisson1", "Categorical1")]

# run RMixtCompt in unsupervised clustering mode + data as matrix
res <- mixtCompLearn(simData$dataLearn$matrix, model, algo, nClass = 2:4)

# plot
plotDiscrimVar(res)
plotDiscrimVar(res$res[[2]])
}

}
\seealso{
\code{\link{computeDiscrimPowerVar}}

Other plot: \code{\link{heatmapClass}},
  \code{\link{heatmapTikSorted}}, \code{\link{heatmapVar}},
  \code{\link{histMisclassif}},
  \code{\link{plot.MixtCompLearn}},
  \code{\link{plot.MixtComp}},
  \code{\link{plotConvergence}}, \code{\link{plotCrit}},
  \code{\link{plotDataBoxplot}}, \code{\link{plotDataCI}},
  \code{\link{plotDiscrimClass}},
  \code{\link{plotParamConvergence}},
  \code{\link{plotProportion}}
}
\author{
Matthieu MARBAC
}
\concept{plot}
