% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT_plotfunctions.R
\name{plotDiscrimVbles}
\alias{plotDiscrimVbles}
\title{Barplot of the discriminative power of the variables}
\usage{
plotDiscrimVbles(output, ylim = c(0, 1), ...)
}
\arguments{
\item{output}{object returned by function \emph{mixtCompCluster}}

\item{ylim}{vector of legnth 2 defining the range of y-axis}

\item{...}{arguments to be passed to plot_ly}
}
\description{
Barplot of the discriminative power of the variables
}
\details{
The discriminative power of variable j is defined by 1 - Cj
Cj=  - sum_{k=1}^K sum_{i=1}^n P(Z_i=k|x_{ij}) ln(P(Z_i=k|x_{ij})) / (n*lnK)
}
\examples{
\donttest{

library(RMixtComp)

data(simData)
 
# define the algorithm's parameters
algo <- createAlgo()

# keep only 3 variables
desc <- simDesc$unsupervised[c("Gaussian1", "Poisson1", "Categorical1")]

# run RMixtCompt in unsupervised clustering mode + data as matrix
res <- mixtCompLearn(simDataLearn$matrix, desc, algo, nClass = 2:4)

# plot
plotDiscrimVbles(res)
plotDiscrimVbles(res$res[[2]])
}

}
\seealso{
\link{computeDiscrimPowerVar}

Other plot: \code{\link{heatmapClass}},
  \code{\link{heatmapTikSorted}},
  \code{\link{heatmapVbles}}, \code{\link{histMisclassif}},
  \code{\link{plotConvergence}},
  \code{\link{plotDataBoxplot}}, \code{\link{plotDataCI}},
  \code{\link{plotDiscrimClass}}
}
\author{
Matthieu MARBAC
}
\concept{plot}
