% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT_plotConvergence.R
\name{plotParamConvergence}
\alias{plotParamConvergence}
\title{Evolution of parameters}
\usage{
plotParamConvergence(output, var, ...)
}
\arguments{
\item{output}{object returned by function \emph{mixtCompLearn}}

\item{var}{name of the variable}

\item{...}{graphical parameters}
}
\description{
Plot the evolution of estimated parameters after the burn-in phase.
}
\examples{
\donttest{
library(RMixtComp)

data(simData)

# define the algorithm's parameters
algo <- createAlgo()

# run RMixtCompt in unsupervised clustering mode + data as matrix
res <- mixtCompLearn(simData$dataLearn$matrix, simData$model$unsupervised, algo, nClass = 2:4)

# plot
plotParamConvergence(res, "Gaussian1")
plotParamConvergence(res, "Poisson1")

}

}
\seealso{
Other plot: \code{\link{heatmapClass}},
  \code{\link{heatmapTikSorted}}, \code{\link{heatmapVar}},
  \code{\link{histMisclassif}},
  \code{\link{plot.MixtCompLearn}},
  \code{\link{plot.MixtComp}},
  \code{\link{plotConvergence}}, \code{\link{plotCrit}},
  \code{\link{plotDataBoxplot}}, \code{\link{plotDataCI}},
  \code{\link{plotDiscrimClass}},
  \code{\link{plotDiscrimVar}},
  \code{\link{plotProportion}}
}
\author{
Quentin Grimonprez
}
\concept{plot}
