% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTCOMP_methods.R
\name{print.MixtComp}
\alias{print.MixtComp}
\title{Print Values}
\usage{
\method{print}{MixtComp}(x, nVarMaxToPrint = 5, ...)
}
\arguments{
\item{x}{\emph{MixtComp} object}

\item{nVarMaxToPrint}{number of variables to display (including z_class)}

\item{...}{Not used.}
}
\description{
Print a \emph{MixtComp} object
}
\examples{
data(iris)

# run RMixtCompt in unsupervised clustering mode and in basic mode
resLearn <- mixtCompLearn(iris[, -5], nClass = 2:4)

print(resLearn$res[[2]])

}
\seealso{
\code{\link{mixtCompLearn}} \code{\link{mixtCompPredict}}
}
