% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTCOMP_methods.R
\name{print.MixtCompLearn}
\alias{print.MixtCompLearn}
\title{Print Values

Print a \emph{MixtCompLearn} object}
\usage{
\method{print}{MixtCompLearn}(x, nVarMaxToPrint = 5, ...)
}
\arguments{
\item{x}{\emph{MixtCompLearn} object}

\item{nVarMaxToPrint}{number of variables to display (including z_class)}

\item{...}{Not used.}
}
\description{
Print Values

Print a \emph{MixtCompLearn} object
}
\examples{
\donttest{
data(simData)
 
# define the algorithm's parameters
algo <- list(nbBurnInIter = 100,
             nbIter = 100,
             nbGibbsBurnInIter = 50,
             nbGibbsIter = 50,
             nInitPerClass = 10,
             nSemTry = 20,
             confidenceLevel = 0.95)

# keep only 3 variables
model <- simDesc$unsupervised[c("Gaussian1", "Poisson1", "Categorical1")]

# run RMixtCompt in unsupervised clustering mode + data as matrix
resLearn <- mixtCompLearn(simDataLearn$matrix, model, algo, nClass = 2:4)

print(resLearn)
}

}
\seealso{
\link{mixtCompLearn} \link{mixtCompPredict}
}
