% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMixtComp-package.R
\docType{package}
\name{RMixtComp-package}
\alias{RMixtComp-package}
\title{RMixtComp}
\description{
Mixture Composer is a project to build mixture models with heterogeneous data
sets. The conditional independence property enables composing all the
data sets in a single mixture model.
}
\details{
\tabular{ll}{
Package: \tab RMixtComp\cr
Type: \tab Package\cr
Version: \tab 0.3\cr
Date: \tab 2018-06-07\cr
}
}
\examples{
\dontrun{
# path to files
pathToData <- system.file("extdata", "data.csv", package = "RMixtComp")
pathToDescriptor <- system.file("extdata", "descUnsupervised.csv", package = "RMixtComp")

resGetData <- getData(c(pathToData, pathToDescriptor))


# define the algorithm's parameters
mcStrategy <- list(nbBurnInIter = 100,
                   nbIter = 100,
                   nbGibbsBurnInIter = 50,
                   nbGibbsIter = 50,
                   nInitPerClass = 10,
                   nSemTry = 20)

# run RMixtCompt for clustering
res <- mixtCompCluster(resGetData$lm, mcStrategy, nbClass = 2, confidenceLevel = 0.95)

# run RMixtCompt for predicting
resPred <- mixtCompPredict(resGetData$lm, res$variable$param, mcStrategy, nbClass = 2, confidenceLevel = 0.95)


}

}
\seealso{
\link{mixtCompCluster} \link{mixtCompPredict}
}
\author{
Vincent KUBICKI, Matthieu MARBAC, Quentin GRIMONPREZ, Serge IOVLEFF

Maintainer: Vincent Kubicki  <vincent.kubicki@inria.fr>
}
\keyword{package}
