% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMixtComp-package.R
\docType{data}
\name{simData}
\alias{simData}
\title{Simulated Heterogeneous data}
\format{Three object: \emph{simDataLearn}, \emph{simDataPredict} and \emph{simDesc}.

\emph{simDataLearn} is a list containing the data in the three accepted format (list, data.frame and matrix). Data consists of 200 individuals and 9 variables.

\emph{simDataPredict} is a list containing the data in the three accepted format (list, data.frame and matrix). Data consists of 100 individuals and 8 variables.

\emph{simDesc} is a list containing the descriptor lists used for clustering \emph{simDesc$unsupervised} and classification \emph{simDesc$supervised}.}
\usage{
data(simData)
}
\description{
Simulated Heterogeneous data
}
\examples{
data(simData)
str(simDataLearn)
str(simDataPredict)
str(simDesc)
  
}
\keyword{data}
