% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTCOMP_methods.R
\name{summary.MixtComp}
\alias{summary.MixtComp}
\title{MixtComp Object Summaries}
\usage{
\method{summary}{MixtComp}(object, ...)
}
\arguments{
\item{object}{\emph{MixtComp} object}

\item{...}{Not used.}
}
\description{
Summary of a \emph{MixtComp} object
}
\examples{
data(iris)

# run RMixtCompt in unsupervised clustering mode and in basic mode
resLearn <- mixtCompLearn(iris[, -5], nClass = 2:4)

summary(resLearn$res[[2]])

}
\seealso{
\code{\link{mixtCompLearn}} \code{\link{print.MixtComp}}
}
\author{
Quentin Grimonprez
}
