% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTCOMP_methods.R
\name{summary.MixtCompLearn}
\alias{summary.MixtCompLearn}
\title{MixtCompLearn Object Summaries

Summary of a \code{\link{MixtCompLearn}} object}
\usage{
\method{summary}{MixtCompLearn}(object, ...)
}
\arguments{
\item{object}{\code{\link{MixtCompLearn}} object}

\item{...}{Not used.}
}
\description{
MixtCompLearn Object Summaries

Summary of a \code{\link{MixtCompLearn}} object
}
\examples{
\donttest{
data(simData)
 
# define the algorithm's parameters
algo <- list(nbBurnInIter = 100,
             nbIter = 100,
             nbGibbsBurnInIter = 50,
             nbGibbsIter = 50,
             nInitPerClass = 10,
             nSemTry = 20,
             confidenceLevel = 0.95)

# run RMixtCompt in unsupervised clustering mode + data as matrix
resLearn <- mixtCompLearn(simDataLearn$matrix, desc$unsupervised, algo, nbClass = 2:4)

summary(resLearn)
}

}
\seealso{
\link{MixtCompLearn} \link{print.MixtCompLearn}
}
