% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTCOMP_misc.R
\name{computeDiscrimPowerVar}
\alias{computeDiscrimPowerVar}
\alias{computeDiscrimPowerClass}
\title{Discriminative power}
\usage{
computeDiscrimPowerVar(outMixtComp)

computeDiscrimPowerClass(outMixtComp)
}
\arguments{
\item{outMixtComp}{object of class \emph{MixtCompLearn} or \emph{MixtComp} obtained using \link{mixtCompLearn} or \link{mixtCompPredict} functions.}
}
\value{
the discriminative power
}
\description{
Compute the discriminative power of each variable or classe
}
\details{
The discriminative power of variable j is defined by 1 - C(j)
\deqn{C(j)=  - \sum_{k=1}^K sum_{i=1}^n P(Z_i=k|x_{ij}) \log(P(Z_i=k|x_{ij})) / (n*\log(K))}

A high value (close to one) means that the variable is highly discriminating. A low value (close to zero) means that the variable is poorly discriminating.


The discriminative power of class k is defined by 1 - D(k)
\deqn{D(k) =  -\sum_{i=1}^n P(Z_i=k|x_i) \log(P(Z_i=k|x_i)) / (n*\exp(-1))}
}
\examples{
\donttest{
data(simData)
 
# define the algorithm's parameters
algo <- list(nbBurnInIter = 100,
             nbIter = 100,
             nbGibbsBurnInIter = 50,
             nbGibbsIter = 50,
             nInitPerClass = 10,
             nSemTry = 20,
             confidenceLevel = 0.95)

# keep only 3 variables
model <- simData$model$unsupervised[c("Gaussian1", "Poisson1", "Categorical1")]

# run RMixtComp for clustering
resLearn <- mixtCompLearn(simData$dataLearn$matrix, model, algo, nClass = 2)

discVar <- computeDiscrimPowerVar(resLearn)
discClass <- computeDiscrimPowerClass(resLearn)

# graphic representation of discVar
plotDiscrimVar(resLearn)
# graphic representation of discClass
plotDiscrimClass(resLearn)
}

}
\seealso{
\code{\link{plotDiscrimClass}} \code{\link{plotDiscrimVar}}
}
\author{
Matthieu Marbac
}
