% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTCOMP_misc.R
\name{computeSimilarityVar}
\alias{computeSimilarityVar}
\alias{computeSimilarityClass}
\title{Similarity}
\usage{
computeSimilarityVar(outMixtComp)

computeSimilarityClass(outMixtComp)
}
\arguments{
\item{outMixtComp}{object of class \emph{MixtCompLearn} or \emph{MixtComp} obtained using \link{mixtCompLearn} or \link{mixtCompPredict} functions.}
}
\value{
a similarity matrix
}
\description{
Compute the similarity between variables (or classes)
}
\details{
The similarities between variables j and h is defined by 1 - Delta(j,h)
Delta(j,h)^2 = (1/n) * sum_{i=1}^n sum_{k=1}^K (P(Z_i=k|x_{ij}) - P(Z_i=k|x_{ih}))^2

The similarities between classes k and g is defined by 1 - Sigma(k,g)
Sigma(k,g)^2 = (1/n) * sum_{i=1}^n (P(Z_i=k|x_i) - P(Z_i=g|x_i))^2
}
\examples{
\donttest{
data(simData)
 
# define the algorithm's parameters
algo <- list(nbBurnInIter = 100,
             nbIter = 100,
             nbGibbsBurnInIter = 50,
             nbGibbsIter = 50,
             nInitPerClass = 10,
             nSemTry = 20,
             confidenceLevel = 0.95)

# keep only 3 variables
desc <- simDesc$unsupervised[c("Gaussian1", "Poisson1", "Categorical1")]

# run RMixtCompt for clustering
resLearn <- mixtCompLearn(simDataLearn$matrix, desc, algo, nClass = 2)

simVar <- computeSimilarityVar(resLearn)
simClass <- computeSimilarityClass(resLearn)
}

}
\seealso{
\link{heatmapVbles} \link{heatmapClass}
}
