% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTCOMP_getData.R
\name{getData}
\alias{getData}
\title{Import data from a csv file}
\usage{
getData(...)
}
\arguments{
\item{...}{a vector containing the names of the two files to import: the data file in first position and the descriptor file in second. Or a list of two matrix: the data in first position and the descriptor in second position. The matrix must contains character element.}
}
\value{
a list containing :
\describe{
  \item{lm}{imported data for \link{mixtCompCluster} and \link{mixtCompPredict}}
  \item{warnlog}{warning message}
}
}
\description{
Import the data file and the descriptor file in the format required by \link{mixtCompCluster} and \link{mixtCompPredict} functions.
}
\details{
There is two ways of loading data: with a data csv file and a descriptor csv file or with two matrix containing the data as characters.


The data file is a csv file (; separator), each column corresponds to a variable. The first row contains the name of each variable.
Others rows correspond to the different samples. See the section associated with each model for more details about data format.


The descriptor file is a csv file (; separator) containing 3 rows. The first row contains the names of the variables.
This first row must be the same as the first row of the data file. The second row contains the model associated with each variables.
Eight models are available in RMixtComp: \emph{Gaussian_sjk}, \emph{Categorical_pjk}, \emph{Poisson_k}, \emph{Weibull}, \emph{Functional}, \emph{FunctionalSharedAlpha}, \emph{Rank}, \emph{Ordinal}. 
The \emph{FunctionalSharedAlpha} is a variant of the Functional model with the alpha parameter shared between clusters. It means that the start and end of each subregression will be the same across the clusters.
The third row is optional and contains parameters associated with each model.
}
\section{Gaussian data}{

Gaussian data are real values with the dot as decimal separator.
Missing data are indicated by a \emph{?}. Partial data can be provided through intervals denoted by 
\emph{[a:b]} where \emph{a} (resp. \emph{b}) is a real or \emph{-inf} (resp. \emph{+inf}).
}

\section{Categorical Data}{

Categorical data must be consecutive integer with 1 as minimal value. Missing data are indicated by a \emph{?}.
For partial data, a list of possible values can be provided by \emph{{a_1,\dots,a_j}}, 
where \emph{a_i} denotes a categorical value.
}

\section{Poisson Data}{

Poisson data must be positive integer. Missing data are indicated by a \emph{?}.
}

\section{Weibull Data}{

Weibull data are real positive values with the dot as decimal separator.
Missing data are indicated by a \emph{?}. Partial data can be provided through intervals denoted by 
\emph{[a:b]} where \emph{a} and \emph{b} are  positive reals.  \emph{b} can be \emph{+inf}.
}

\section{Ordinal Data}{

Ordinal data must be consecutive integer with 0 as minimal value. Missing data are indicated by a \emph{?}.
For partial data, a list of possible values can be provided by \emph{\{a_1,\dots,a_j\}}, 
where \emph{a_i} denotes a categorical value.
}

\section{Rank data}{

The format of a rank is: \emph{o_1, \dots, o_j} where o_1 is an integer corresponding to the the number of the object ranked in 1st position.
For example: 4,2,1,3 means that the fourth object is ranked first then the second object is in second position and so on.
Missing data can be specified by replacing and object by a \emph{?} or a list of potential object, for example: \emph{4, \{2 3\}, \{2 1\}, ?} means that 
the object ranked in second position is either the object number 2 or the object number 3, then the object ranked in third position is either the object 2 or 1 and the last one can be anything.
A totally missing rank is spedified by \emph{?,?,\dots,?}
}

\section{Functional data}{

The format of a fonctional data is: \emph{time_1:value_1,\dots, time_j:value_j}. Between individuals, functional data can have different length and different time.
In the case of a functional model, \emph{nSub: i, nCoeff: k} must be indicated in the third row of the descriptor file.
\emph{i} is the number of subregressions in a functional data and \emph{k} the number of coefficients
 of each regression (2 = linear, 3 = quadratic, ...). Missing data are not supported.
}

\section{(semi-)supervised clustering}{

To perform a (semi-)supervised clustering, user can add a variable names \emph{z_class} in the data and descriptor files with \emph{LatentClass} as model in the descriptor file.
}

\examples{
\dontrun{
# Case with csv file
pathToData <- system.file("extdata", "data.csv", package = "RMixtComp")
pathToDescriptor <- system.file("extdata", "descUnsupervised.csv", package = "RMixtComp")

out <- getData(c(pathToData, pathToDescriptor)) 

# Case with R object
var <- list()
var$z_class <- zParam()
var$Gaussian1 <- gaussianParam("Gaussian1")
var$Categorical1 <- categoricalParam("Categorical1")

dat <- dataGenerator(100, 0.8, var) 

out <- getData(list(dat$data, dat$descriptor)) 
}

}
\seealso{
\link{mixtCompCluster} \link{mixtCompPredict}
}
\author{
Vincent Kubicki
}
