% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{getType}
\alias{getType}
\title{Get the type of model for each variable}
\usage{
getType(outMixtcomp, with.z_class = FALSE)
}
\arguments{
\item{with.z_class}{if TRUE, the type of z_class is returned.}

\item{outMixtComp}{output object of \link{mixtCompCluster} or \link{mixtCompPredict} fucntions.}
}
\value{
a vector containing the type of models associated with each individual.
}
\description{
Get the type of model for each variable
}
\examples{
\dontrun{
# load the data
resGetData <- getData(c("dataTrain.csv", "descriptor.csv")) 
resGetNewData <- getData(c("dataTest.csv", "descriptor.csv")) 

# define the algorithm's parameters
mcStrategy <- list(nbBurnInIter = 100,
nbIter = 100,
nbGibbsBurnInIter = 50,
nbGibbsIter = 50,
parameterEdgeAuthorized = FALSE)

# run RMixtCompt for clustering
res <- mixtCompCluster(resGetData$lm, mcStrategy, nbClass = 2, confidenceLevel = 0.95)

# get type
type <- getType(res)
}

}

