% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTCOMP_getter.R
\name{getType}
\alias{getType}
\alias{getModel}
\alias{getVarNames}
\title{Names and Types Getters}
\usage{
getType(outMixtComp, with.z_class = FALSE)

getModel(outMixtComp, with.z_class = FALSE)

getVarNames(outMixtComp, with.z_class = FALSE)
}
\arguments{
\item{outMixtComp}{object of class \emph{MixtCompLearn} or \emph{MixtComp} obtained using \link{mixtCompLearn} or \link{mixtCompPredict} functions.}

\item{with.z_class}{if TRUE, the type of z_class is returned.}
}
\value{
a vector containing the type of models, names associated with each individual.
}
\description{
getType returns the type output of a MixtComp object, getModel returns the model object, getVarNames returns the name for each variable
}
\examples{
\donttest{
data(simData)
 
# define the algorithm's parameters
algo <- list(nbBurnInIter = 100,
             nbIter = 100,
             nbGibbsBurnInIter = 50,
             nbGibbsIter = 50,
             nInitPerClass = 10,
             nSemTry = 20,
             confidenceLevel = 0.95)

# run RMixtComp for clustering
resLearn <- mixtCompLearn(simData$dataLearn$matrix, simData$model$unsupervised, algo, nClass = 2)

# get type
type <- getType(resLearn)

# get model object
model <- getModel(resLearn)

# get variable names
varNames <- getVarNames(resLearn)

}

}
\seealso{
Other getter: \code{\link{getBIC}},
  \code{\link{getCompletedData}},
  \code{\link{getEmpiricTik}}, \code{\link{getParam}},
  \code{\link{getPartition}}
}
\author{
Quentin Grimonprez
}
\concept{getter}
