% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTCOMP_plotfunctions.R
\name{heatmapClass}
\alias{heatmapClass}
\title{Heatmap of the similarities between classes about clustering}
\usage{
heatmapClass(output)
}
\arguments{
\item{output}{object returned by function \link{mixtCompCluster}}
}
\details{
The similarities between classes k and g is defined by 1 - Sigma(k,g)
Sigma(k,g)^2 = (1/n) * sum_{i=1}^n (P(Z_i=k|x_i) - P(Z_i=g|x_i))^2
}
\examples{
\dontrun{
# path to files
pathToData <- system.file("extdata", "data.csv", package = "RMixtComp")
pathToDescriptor <- system.file("extdata", "descriptor.csv", package = "RMixtComp")

resGetData <- getData(c(pathToData, pathToDescriptor))


# define the algorithm's parameters
mcStrategy <- list(nbBurnInIter = 100,
                   nbIter = 100,
                   nbGibbsBurnInIter = 50,
                   nbGibbsIter = 50)

# run RMixtCompt for clustering
res <- mixtCompCluster(resGetData$lm, mcStrategy, nbClass = 2, confidenceLevel = 0.95)

# plot
heatmapClass(res)

} 

}
\author{
Matthieu MARBAC
}
