% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT_plotfunctions.R
\name{heatmapClass}
\alias{heatmapClass}
\title{Heatmap of the similarities between classes about clustering}
\usage{
heatmapClass(output, pkg = c("ggplot2", "plotly"), ...)
}
\arguments{
\item{output}{object returned by function \emph{mixtCompLearn}}

\item{pkg}{"ggplot2" or "plotly". Package used to plot}

\item{...}{arguments to be passed to plot_ly. For pkg = "ggplot2", addValues = TRUE prints similarity values on the heatmap}
}
\description{
Heatmap of the similarities between classes about clustering
}
\details{
The similarities between classes k and g is defined by 1 - Sigma(k,g)
Sigma(k,g)^2 = (1/n) * sum_{i=1}^n (P(Z_i=k|x_i) - P(Z_i=g|x_i))^2
}
\examples{
\donttest{
library(RMixtComp)

data(simData)
 
# define the algorithm's parameters
algo <- createAlgo()

# keep only 3 variables
desc <- simDesc$unsupervised[c("Gaussian1", "Poisson1", "Categorical1")]

# run RMixtCompt in unsupervised clustering mode + data as matrix
res <- mixtCompLearn(simDataLearn$matrix, desc, algo, nClass = 2:4)
# plot
heatmapClass(res)
heatmapClass(res$res[[2]])
} 

}
\seealso{
\link{computeSimilarityClass}

Other plot: \code{\link{heatmapTikSorted}},
  \code{\link{heatmapVar}}, \code{\link{histMisclassif}},
  \code{\link{plot.MixtComp}},
  \code{\link{plotConvergence}},
  \code{\link{plotDataBoxplot}}, \code{\link{plotDataCI}},
  \code{\link{plotDiscrimClass}},
  \code{\link{plotDiscrimVar}},
  \code{\link{plotProportion}}
}
\author{
Matthieu MARBAC
}
\concept{plot}
