% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT_heatmap.R
\name{heatmapTikSorted}
\alias{heatmapTikSorted}
\title{Heatmap of the tik = P(Z_i=k|x_i)}
\usage{
heatmapTikSorted(output, pkg = c("ggplot2", "plotly"), ...)
}
\arguments{
\item{output}{object returned by function \emph{mixtCompLearn}}

\item{pkg}{"ggplot2" or "plotly". Package used to plot}

\item{...}{arguments to be passed to plot_ly}
}
\description{
Heatmap of the tik = P(Z_i=k|x_i)
}
\details{
Observation are sorted according to the hard partition then for each component
they are sorted by decreasing order of their tik's
}
\examples{
\donttest{
library(RMixtComp)

data(simData)
 
# define the algorithm's parameters
algo <- createAlgo()

# keep only 3 variables
model <- simDesc$unsupervised[c("Gaussian1", "Poisson1", "Categorical1")]

# run RMixtCompt in unsupervised clustering mode + data as matrix
res <- mixtCompLearn(simDataLearn$matrix, model, algo, nClass = 2:4)

# plot
heatmapTikSorted(res)
heatmapTikSorted(res$res[[2]])
}

}
\seealso{
getTik

Other plot: \code{\link{heatmapClass}},
  \code{\link{heatmapVar}}, \code{\link{histMisclassif}},
  \code{\link{plot.MixtCompLearn}},
  \code{\link{plot.MixtComp}},
  \code{\link{plotConvergence}}, \code{\link{plotCrit}},
  \code{\link{plotDataBoxplot}}, \code{\link{plotDataCI}},
  \code{\link{plotDiscrimClass}},
  \code{\link{plotDiscrimVar}},
  \code{\link{plotProportion}}
}
\author{
Matthieu MARBAC
}
\concept{plot}
