% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT_plotfunctions.R
\name{heatmapVbles}
\alias{heatmapVbles}
\title{Heatmap of the similarities between variables about clustering}
\usage{
heatmapVbles(output, pkg = c("ggplot2", "plotly"), ...)
}
\arguments{
\item{output}{object returned by function \emph{mixtCompCluster}}

\item{pkg}{"ggplot2" or "plotly". Package used to plot}

\item{...}{arguments to be passed to plot_ly. For pkg = "ggplot2", addValues = TRUE prints similarity values on the heatmap}
}
\description{
Heatmap of the similarities between variables about clustering
}
\details{
The similarities between variables j and h is defined by 1 - Delta(j,h)
Delta(j,h)^2 = (1/n) * sum_{i=1}^n sum_{k=1}^K (P(Z_i=k|x_{ij}) - P(Z_i=k|x_{ih}))^2
}
\examples{
\donttest{
library(RMixtComp)

data(simData)
 
# define the algorithm's parameters
algo <- createAlgo()

# keep only 3 variables
desc <- simDesc$unsupervised[c("Gaussian1", "Poisson1", "Categorical1")]

# run RMixtCompt in unsupervised clustering mode + data as matrix
res <- mixtCompLearn(simDataLearn$matrix, desc, algo, nClass = 2:4)

# plot
heatmapVbles(res)
heatmapVbles(res$res[[2]])
}

}
\seealso{
\link{computeSimilarityVar}

Other plot: \code{\link{heatmapClass}},
  \code{\link{heatmapTikSorted}},
  \code{\link{histMisclassif}},
  \code{\link{plotConvergence}},
  \code{\link{plotDataBoxplot}}, \code{\link{plotDataCI}},
  \code{\link{plotDiscrimClass}},
  \code{\link{plotDiscrimVar}}
}
\author{
Matthieu MARBAC
}
\concept{plot}
