% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT_plotfunctions.R
\name{histMisclassif}
\alias{histMisclassif}
\title{Histogram of the misclassification probabilities}
\usage{
histMisclassif(output, ...)
}
\arguments{
\item{output}{object returned by function \emph{mixtCompCluster}}

\item{...}{arguments to be passed to plot_ly}
}
\description{
Histogram of the misclassification probabilities
}
\details{
Missclassification probability of observation i is denoted err_i
err_i = 1 - max_{k={1,\ldots,K}} P(Z_i=k|x_i)
Histgrams of err_i's can be plot for a specific class, all classes or every class
}
\examples{
\donttest{
library(RMixtComp)

data(simData)
 
# define the algorithm's parameters
algo <- createAlgo()

# keep only 3 variables
desc <- simDesc$unsupervised[c("Gaussian1", "Poisson1", "Categorical1")]

# run RMixtCompt in unsupervised clustering mode + data as matrix
res <- mixtCompLearn(simDataLearn$matrix, desc, algo, nClass = 2:4)

# plot
histMisclassif(res)
histMisclassif(res$res[[2]])
}

}
\seealso{
Other plot: \code{\link{heatmapClass}},
  \code{\link{heatmapTikSorted}}, \code{\link{heatmapVar}},
  \code{\link{plotConvergence}},
  \code{\link{plotDataBoxplot}}, \code{\link{plotDataCI}},
  \code{\link{plotDiscrimClass}},
  \code{\link{plotDiscrimVar}}
}
\author{
Matthieu MARBAC
}
\concept{plot}
