% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTCOMP_methods.R
\name{plot.MixtCompLearn}
\alias{plot.MixtCompLearn}
\title{Plot of a \emph{MixtCompLearn} object}
\usage{
\method{plot}{MixtCompLearn}(x, nVarMaxToPlot = 5, pkg = c("ggplot2",
  "plotly"), ...)
}
\arguments{
\item{x}{\emph{MixtCompLearn} object}

\item{nVarMaxToPlot}{number of variables to display}

\item{pkg}{"ggplot2" or "plotly". Package used to plot}

\item{...}{Not used.}
}
\description{
Plot of a \emph{MixtCompLearn} object
}
\examples{
\donttest{
data(simData)
 
# define the algorithm's parameters
algo <- list(nbBurnInIter = 100,
             nbIter = 100,
             nbGibbsBurnInIter = 50,
             nbGibbsIter = 50,
             nInitPerClass = 10,
             nSemTry = 20,
             confidenceLevel = 0.95)

# keep only 3 variables
model <- simDesc$unsupervised[c("Gaussian1", "Poisson1", "Categorical1")]

# run RMixtCompt in unsupervised clustering mode + data as matrix
resLearn <- mixtCompLearn(simDataLearn$matrix, model, algo, nClass = 2:4)

plot(resLearn)
}
}
\seealso{
\link{mixtCompLearn} \link{mixtCompPredict}

Other plot: \code{\link{heatmapClass}},
  \code{\link{heatmapTikSorted}}, \code{\link{heatmapVar}},
  \code{\link{histMisclassif}},
  \code{\link{plot.MixtComp}},
  \code{\link{plotConvergence}}, \code{\link{plotCrit}},
  \code{\link{plotDataBoxplot}}, \code{\link{plotDataCI}},
  \code{\link{plotDiscrimClass}},
  \code{\link{plotDiscrimVar}},
  \code{\link{plotProportion}}
}
\concept{plot}
