% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT_plotConvergence.R
\name{plotConvergence}
\alias{plotConvergence}
\title{Convergence of algorithm}
\usage{
plotConvergence(output, ...)
}
\arguments{
\item{output}{object returned by function \emph{mixtCompLearn}}

\item{...}{graphical parameters}
}
\description{
Plot the evolution of the completed loglikelihood during the SEM algorithm. The vertical line denotes the end of the burn-in phase.
}
\details{
This function can be used to check the convergence and choose the parameters nbBurnInIter and nbIter from mcStrategy.
}
\examples{
\donttest{
library(RMixtComp)

data(simData)
 
# define the algorithm's parameters
algo <- createAlgo()

# keep only 3 variables
desc <- simDesc$unsupervised[c("Gaussian1", "Poisson1", "Categorical1")]

# run RMixtCompt in unsupervised clustering mode + data as matrix
res <- mixtCompLearn(simDataLearn$matrix, desc, algo, nClass = 2:4)
# plot
plotConvergence(res)
plotConvergence(res$res[[2]])
} 

}
\seealso{
Other plot: \code{\link{heatmapClass}},
  \code{\link{heatmapTikSorted}}, \code{\link{heatmapVar}},
  \code{\link{histMisclassif}},
  \code{\link{plot.MixtComp}},
  \code{\link{plotDataBoxplot}}, \code{\link{plotDataCI}},
  \code{\link{plotDiscrimClass}},
  \code{\link{plotDiscrimVar}},
  \code{\link{plotProportion}}
}
\author{
Quentin Grimonprez
}
\concept{plot}
