% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLOT_plotfunctions.R
\name{plotDiscrimClass}
\alias{plotDiscrimClass}
\title{Barplot of the discriminative power of the classes}
\usage{
plotDiscrimClass(output, ylim = c(0, 1), pkg = c("ggplot2", "plotly"),
  ...)
}
\arguments{
\item{output}{object returned by function \emph{mixtCompCluster}}

\item{ylim}{vector of length 2 defining the range of y-axis}

\item{pkg}{"ggplot2" or "plotly". Package used to plot}

\item{...}{arguments to be passed to plot_ly}
}
\description{
Barplot of the discriminative power of the classes
}
\details{
The discriminative power of class k is defined by 1 - Dk
Dk =  - sum_{i=1}^n P(Z_i=k|x_i) ln(P(Z_i=k|x_i)) / (n*exp(-1))
}
\examples{
\donttest{
library(RMixtComp)

data(simData)
 
# define the algorithm's parameters
algo <- createAlgo()

# keep only 3 variables
desc <- simDesc$unsupervised[c("Gaussian1", "Poisson1", "Categorical1")]

# run RMixtCompt in unsupervised clustering mode + data as matrix
res <- mixtCompLearn(simDataLearn$matrix, desc, algo, nClass = 2:4)

# plot
plotDiscrimClass(res)
plotDiscrimClass(res$res[[2]])
}

}
\seealso{
\link{computeDiscrimPowerClass}

Other plot: \code{\link{heatmapClass}},
  \code{\link{heatmapTikSorted}},
  \code{\link{heatmapVbles}}, \code{\link{histMisclassif}},
  \code{\link{plotConvergence}},
  \code{\link{plotDataBoxplot}}, \code{\link{plotDataCI}},
  \code{\link{plotDiscrimVbles}}
}
\author{
Matthieu MARBAC
}
\concept{plot}
