% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTCOMP_plotfunctions.R
\name{plotDiscrimVbles}
\alias{plotDiscrimVbles}
\title{Barplot of the discriminative power of the variables}
\usage{
plotDiscrimVbles(output, ylim = c(0, 1))
}
\arguments{
\item{output}{object returned by function \link{mixtCompCluster}}

\item{ylim}{vector of legnth 2 defining the range of y-axis}
}
\details{
The discriminative power of variable j is defined by 1 - Cj
Cj=  - sum_{k=1}^K sum_{i=1}^n P(Z_i=k|x_{ij}) ln(P(Z_i=k|x_{ij})) / (n*lnK)
}
\examples{
\dontrun{
# path to files
pathToData <- system.file("extdata", "data.csv", package = "RMixtComp")
pathToDescriptor <- system.file("extdata", "descriptor.csv", package = "RMixtComp")

resGetData <- getData(c(pathToData, pathToDescriptor))


# define the algorithm's parameters
mcStrategy <- list(nbBurnInIter = 100,
                   nbIter = 100,
                   nbGibbsBurnInIter = 50,
                   nbGibbsIter = 50,
                   ratioInitialization = 0.1)

# run RMixtCompt for clustering
res <- mixtCompCluster(resGetData$lm, mcStrategy, nbClass = 2, confidenceLevel = 0.95)

# plot
plotDiscrimVbles(res)

}

}
\author{
Matthieu MARBAC
}
