% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTCOMP_methods.R
\name{print.MixtCompLearn}
\alias{print.MixtCompLearn}
\title{Print Values

Print a \code{\link{MixtCompLearn}} object}
\usage{
\method{print}{MixtCompLearn}(x, nVarMaxToPrint = 5, ...)
}
\arguments{
\item{x}{\code{\link{MixtCompLearn}} object}

\item{nVarMaxToPrint}{number of varaibles to display (including z_class)}

\item{...}{Not used.}
}
\description{
Print Values

Print a \code{\link{MixtCompLearn}} object
}
\examples{
\donttest{
data(simData)
 
# define the algorithm's parameters
algo <- list(nbBurnInIter = 100,
             nbIter = 100,
             nbGibbsBurnInIter = 50,
             nbGibbsIter = 50,
             nInitPerClass = 10,
             nSemTry = 20,
             confidenceLevel = 0.95)

# keep only 3 variables
desc <- simDesc$unsupervised[c("Gaussian1", "Poisson1", "Categorical1")]

# run RMixtCompt in unsupervised clustering mode + data as matrix
resLearn <- mixtCompLearn(simDataLearn$matrix, desc, algo, nClass = 2:4)

print(resLearn)
}

}
\seealso{
\link{mixtCompLearn} \link{mixtCompPredict}
}
