% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTCOMP_methods.R
\name{print.MixtCompLearn}
\alias{print.MixtCompLearn}
\title{Print Values}
\usage{
\method{print}{MixtCompLearn}(x, nVarMaxToPrint = 5, nClass = NULL, ...)
}
\arguments{
\item{x}{\emph{MixtCompLearn} object}

\item{nVarMaxToPrint}{number of variables to display (including z_class)}

\item{nClass}{number of classes of the model to print}

\item{...}{Not used.}
}
\description{
Print a \emph{MixtCompLearn} object
}
\examples{
data(iris)

# run RMixtComp in unsupervised clustering mode and in basic mode
resLearn <- mixtCompLearn(iris[, -5], nClass = 2:4)

print(resLearn)
print(resLearn, nClass = 3)


}
\seealso{
\code{\link{mixtCompLearn}} \code{\link{mixtCompPredict}}
}
\author{
Quentin Grimonprez
}
