% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMixtComp-package.R
\docType{package}
\name{RMixtComp-package}
\alias{RMixtComp-package}
\title{RMixtComp}
\description{
Mixture Composer is a project to build mixture models with heterogeneous data
sets. The conditional independence property enable composing all the
data sets in a single mixture model.
}
\details{
\tabular{ll}{
Package: \tab RMixtComp\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2016-10-10\cr
License: \tab GPL (>=2) \cr
}
}
\examples{
\dontrun{
# load the train data
resGetData <- getData(c("dataTrain.csv", "descriptor.csv")) 
# load the test data
resGetNewData <- getData(c("dataTest.csv", "descriptor.csv")) 

# define the algorithm's parameters
mcStrategy <- list(nbBurnInIter = 100,
nbIter = 100,
nbGibbsBurnInIter = 50,
nbGibbsIter = 50,
parameterEdgeAuthorized = FALSE)

# run RMixtCompt for clustering
res <- mixtCompCluster(resGetData$lm, mcStrategy, nbClass = 2, confidenceLevel = 0.95)

# run RMixtCompt for predicting
resPred <- mixtCompCluster(resGetNewData$lm, res$variable$param, mcStrategy, nbClass = 2, confidenceLevel = 0.95)
}

}
\seealso{
\link{mixtCompCluster} \link{mixtCompPredict}
}
\author{
Vincent Kubicki

Maintainer: Vincent Kubicki  <vincent.kubicki@inria.fr>
}
\keyword{package}
