% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTCOMP_methods.R
\name{summary.MixtComp}
\alias{summary.MixtComp}
\title{MixtComp Object Summaries

Summary of a \code{\link{MixtComp}} object}
\usage{
\method{summary}{MixtComp}(object, ...)
}
\arguments{
\item{object}{\code{\link{MixtComp}} object}

\item{...}{Not used.}
}
\description{
MixtComp Object Summaries

Summary of a \code{\link{MixtComp}} object
}
\examples{
\donttest{
data(simData)
 
# define the algorithm's parameters
algo <- list(nbBurnInIter = 100,
             nbIter = 100,
             nbGibbsBurnInIter = 50,
             nbGibbsIter = 50,
             nInitPerClass = 10,
             nSemTry = 20,
             confidenceLevel = 0.95)

# keep only 3 variables
desc <- simDesc$unsupervised[c("Gaussian1", "Poisson1", "Categorical1")]

# run RMixtCompt in unsupervised clustering mode + data as matrix
resLearn <- mixtCompLearn(simDataLearn$matrix, desc, algo, nClass = 2:4)

summary(resLearn$res[[1]])

}

}
\seealso{
\link{MixtCompLearn} \link{print.MixtComp}
}
