% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MIXTCOMP_methods.R
\name{summary.MixtComp}
\alias{summary.MixtComp}
\title{MixtComp Object Summaries

Summary of a \emph{MixtComp} object}
\usage{
\method{summary}{MixtComp}(object, ...)
}
\arguments{
\item{object}{\emph{MixtComp} object}

\item{...}{Not used.}
}
\description{
MixtComp Object Summaries

Summary of a \emph{MixtComp} object
}
\examples{
\donttest{
data(simData)
 
# define the algorithm's parameters
algo <- list(nbBurnInIter = 100,
             nbIter = 100,
             nbGibbsBurnInIter = 50,
             nbGibbsIter = 50,
             nInitPerClass = 10,
             nSemTry = 20,
             confidenceLevel = 0.95)

# keep only 3 variables
model <- simData$model$unsupervised[c("Gaussian1", "Poisson1", "Categorical1")]

# run RMixtCompt in unsupervised clustering mode + data as matrix
resLearn <- mixtCompLearn(simData$dataLearn$matrix, model, algo, nClass = 2:4)

summary(resLearn$res[[1]])

}

}
\seealso{
\code{\link{mixtCompLearn}} \code{\link{print.MixtComp}}
}
